fix: removing unused table components that are replaced, children prop fixes, autotable structure fix
