+++ packages/table/src/table-row.tsx
@@ -1,13 +1,5 @@
 import React from 'react';
-import { DefaultProps, cx, getValidChildren } from '@sk-web-gui/utils';
-
-interface ITableRowProps extends DefaultProps {
-  children: JSX.Element | JSX.Element[] | string;
-}
-
-export interface TableRowProps
-  extends Omit<React.HTMLAttributes<HTMLTableRowElement>, 'color' | 'children'>,
-    ITableRowProps {}
+import { cx, getValidChildren } from '@sk-web-gui/utils';
 
 export const TableRow = React.forwardRef<HTMLTableRowElement, React.ComponentPropsWithoutRef<'tr'>>((props, ref) => {
   const { children, className, ...rest } = props;
