+++ packages/table/src/table.tsx
@@ -3,17 +3,13 @@ import React from 'react';
 import { TableFooter } from './table-footer';
 
 interface UseTableProps {
-  summary?: string;
   background?: boolean;
 }
 
-export interface TableComponentProps
-  extends DefaultProps,
-    UseTableProps,
-    Omit<React.ComponentPropsWithRef<'table'>, 'color'> {}
+export interface TableComponentProps extends DefaultProps, UseTableProps, React.ComponentPropsWithRef<'table'> {}
 
 export const TableComponent = React.forwardRef<HTMLTableElement, TableComponentProps>((props, ref) => {
-  const { summary, background = false, className, children, ...rest } = props;
+  const { background = false, className, children, ...rest } = props;
 
   //MANUEL TABLE
   const validChildren = getValidChildren(children);
@@ -27,9 +23,13 @@ export const TableComponent = React.forwardRef<HTMLTableElement, TableComponentP
 
   const footerItem = validChildren.filter((child) => child.type === TableFooter);
   return (
-    <div className={cx('sk-table-wrapper', className)} data-background={background}>
+    <div
+      className={cx('sk-table-wrapper', className)}
+      data-footer={validChildren.find((child) => child.type === TableFooter) ? true : false}
+      data-background={background}
+    >
       <div className="sk-table-wrapper-inside">
-        <table ref={ref} {...rest} className={'sk-table'} summary={summary ? summary : undefined}>
+        <table ref={ref} {...rest} className={'sk-table'}>
           {tableItems}
         </table>
